refactor: move runtime_requirements / network_requirements into the mixin
ovos-core deprecated runtime_requirements (network_requirements has
been deprecated for years). Move both classproperties into
_LegacyResourcesMixin so the whole 'pre-spec-tools' surface of
OVOSSkill — resources, dialog_renderer, find_resource,
runtime_requirements, network_requirements — disappears as one
removal: drop the mixin from the bases and delete the file.

Subclass overrides keep working: skill_launcher accesses the property
via hasattr/MRO so it does not care whether it lives on OVOSSkill
itself or on a base.
